+++ apps/pooly/config/config.exs
@@ -1,0 +1,1 @@
+use Mix.Config
