+++ apps/pooly/lib/sample_worker.ex
@@ -1,4 +1,4 @@
-defmodule Little.SampleWorker do
+defmodule Pooly.SampleWorker do
   use GenServer
 
   def start_link(_) do
