+++ apps/pooly/test/test_helper.exs
@@ -1,0 +1,1 @@
+ExUnit.start()
