+++ apps/pooly/test/worker_supervisor_test.exs
@@ -1,0 +1,16 @@
+defmodule WorkerSupervisorTest do
+  use ExUnit.Case
+  use Supervisor
+
+  alias Pooly.WorkerSupervisor
+  alias Pooly.SampleWorker
+
+  describe "start_link/1" do
+    test "can start a Supervisor with Process" do
+      {:ok, worker_sup} = WorkerSupervisor.start_link({SampleWorker, :start_link, []})
+
+      {result, _value} = Supervisor.start_child(worker_sup, [[]])
+      assert result === :ok
+    end
+  end
+end
