+++ mix.exs
@@ -1,35 +1,14 @@
-defmodule Little.Mixfile do
+defmodule LittleOtp.Mixfile do
   use Mix.Project
 
   def project do
-    [app: :little,
-     version: "0.1.0",
-     elixir: "~> 1.4",
+    [apps_path: "apps",
      build_embedded: Mix.env == :prod,
      start_permanent: Mix.env == :prod,
      deps: deps()]
   end
 
-  # Configuration for the OTP application
-  #
-  # Type "mix help compile.app" for more information
-  def application do
-    # Specify extra applications you'll use from Erlang/Elixir
-    [extra_applications: [:logger, :httpoison],
-     mod: {Little.Pooly, []}]
-  end
-
-  # Dependencies can be Hex packages:
-  #
-  #   {:my_dep, "~> 0.3.0"}
-  #
-  # Or git/path repositories:
-  #
-  #   {:my_dep, git: "https://github.com/elixir-lang/my_dep.git", tag: "0.1.0"}
-  #
-  # Type "mix help deps" for more examples and options
   defp deps do
-    [{:json, "~> 1.0"},
-    {:httpoison, "~> 0.11.1"}]
+    []
   end
 end
